the inject code works, but injecting dlopen at %rip gives back zero. %rip seems to be in libc so it's likely that the program is stopped at some kind of system call and that dlopen won't work in this context

+++ inject.h
@@ -17,7 +17,7 @@
 int inject_and_run_text (int pid, int size, const uint8_t* text);
 
 
-int inject_dlopen (int pid, const char* szsharedlib, uint8_t flags);
+int inject_dlopen (int pid, const char* szsharedlib, uint32_t flags);
